Extractor function that ignores the x-axis and applies statistics over the y-values.

Extractor function that ignores the x-axis and applies statistics over the y-values.

extract_ignore_x(data)

Arguments

data
Provided internally, users don't need to provide this.

Value

A 1-D numeric vector of y-values from the plot

Details

For example this might be relevant for a mean per-base fastq quality score. This will let you then calculate the overall mean quality of the reads.

See also

Examples

report <- load_multiqc(
   system.file("extdata", "HG00096/multiqc_data.json", package = "TidyMultiqc"),
   sections = "plots",
   plot_opts = list(
     `fastqc_per_base_sequence_quality_plot` = list(
       extractor = extract_ignore_x,
       summary = list(mean = mean),
       prefix = "quality"
     )
   )
)
#> Error in load_multiqc(system.file("extdata", "HG00096/multiqc_data.json",     package = "TidyMultiqc"), sections = "plots", plot_opts = list(fastqc_per_base_sequence_quality_plot = list(extractor = extract_ignore_x,     summary = list(mean = mean), prefix = "quality"))): unused argument (plot_opts = list(fastqc_per_base_sequence_quality_plot = list(extractor = extract_ignore_x, summary = list(mean = mean), prefix = "quality")))