Loads one or more MultiQCs report into a data frame

Loads one or more MultiQCs report into a data frame

load_multiqc(
  paths,
  plots = NULL,
  find_metadata = function(...) {     list() },
  plot_parsers = list(),
  sections = "general"
)

Arguments

paths
A vector of filepaths to multiqc_data.json files

find_metadata
A function that will be called with a sample name and the parsed JSON and returns a named list of metadata fields for the sample

sections
Vector of the sections to include in the output: 'plots' in the list means parse plot data, 'general' means parse the general stats section, and 'raw' means parse the raw data section. This defaults to 'general', which tends to contain the most useful statistics

plot_opts
A named list mapping the internal MultiQC plot name, e.g. "fastqc_per_sequence_quality_scores_plot" to a list of options for that plot. The list can have the following keys:

extractor

Mandatory for scatter/line plots, ignored for bar graphs. A function which converts the raw plot JSON into a some kind of data, usually a vector. Often you will want to use a built-in extract_x functions provided by this package

summary

A named list of functions that each map the output from the extractor function (usually a 1-D vector) to a scalar, to "summarise" it. For example, you might want to use the base::mean() function to summarise the plot. See also the summary_x functions in this package.

prefix

Optional. A new name for this plot. MultiQC sometimes has some unwieldy names for its plot, so this lets you rename them

Value

A tibble (data.frame subclass) with QC data and metadata as columns, and samples as rows

Examples

load_multiqc(
  system.file("extdata", "wgs/multiqc_data.json", package = "TidyMultiqc"),
  sections = c("plots", "general", "raw"),
  plot_opts = list(
    fastqc_per_sequence_quality_scores_plot = list(
      summary = list(`%q30` = summary_q30),
      extractor = extract_histogram,
      prefix = "quality"
    )
  )
)
#> Error in load_multiqc(system.file("extdata", "wgs/multiqc_data.json",     package = "TidyMultiqc"), sections = c("plots", "general",     "raw"), plot_opts = list(fastqc_per_sequence_quality_scores_plot = list(summary = list(`%q30` = summary_q30),     extractor = extract_histogram, prefix = "quality"))): unused argument (plot_opts = list(fastqc_per_sequence_quality_scores_plot = list(summary = list(`%q30` = summary_q30), extractor = extract_histogram, prefix = "quality")))