List the plot identifiers of all the plots in a given multiqc report

List the plot identifiers of all the plots in a given multiqc report

list_plots(path)

Arguments

path
The file path to the multiqc report. This should be a length 1 character vector

Value

A data frame containing n rows, where n is the number of plots in the report you have provided, and two columns:

id

The identifier for the plot. This is the one you should use as a name in plot_opts.

name

The plot title. This is likely what you see in the multiqc report when you open it with your browser.

Details

This is a useful function, because the plot_opts list used in the main load_multiqc function requires these identifiers as names. Refer to the "Plot Extraction" vignette for more information.

Examples

# Ignore this, choose your own filepath as the `filepath` variable
filepath <- system.file("extdata", "HG00096/multiqc_data.json", package = "TidyMultiqc")
# This is the actual invocation
list_plots(filepath)
#> # A tibble: 6 × 2
#>   id                                      title                              
#>   <chr>                                   <chr>                              
#> 1 fastqc_sequence_counts_plot             FastQC: Sequence Counts            
#> 2 fastqc_per_base_sequence_quality_plot   FastQC: Mean Quality Scores        
#> 3 fastqc_per_sequence_quality_scores_plot FastQC: Per Sequence Quality Scores
#> 4 fastqc_per_sequence_gc_content_plot     FastQC: Per Sequence GC Content    
#> 5 fastqc_per_base_n_content_plot          FastQC: Per Base N Content         
#> 6 fastqc_sequence_duplication_levels_plot FastQC: Sequence Duplication Levels